server/requirements.txt
# Core dependencies
fastapi>=0.95.0
uvicorn>=0.21.0
pydantic>=2.0.0
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
httpx>=0.24.0

# Testing
pytest>=7.3.1
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# AI/ML dependencies
dspy-ai>=0.11.0
transformers>=4.35.0
torch>=2.0.0
accelerate>=0.21.0
bitsandbytes>=0.41.0  # For 4-bit quantization
safetensors>=0.3.3

# Database
sqlalchemy>=2.0.0
alembic>=1.11.0
psycopg2-binary>=2.9.6

# Vector store
chromadb>=0.4.0

# Monitoring and logging
prometheus-client>=0.17.0
python-json-logger>=2.0.7

# Utilities
python-multipart>=0.0.6
aiofiles>=23.1.0